Julia functions

Dynare.forecasting! — Function

forecasting!(; periods::Integer,
               forecast_mode::ForecastModes,
               context::Context=context,
               datafile::String="",
               first_obs::PeriodsSinceEpoch=Undated(typemin(Int)),
               first_period::PeriodsSinceEpoch=Undated(0),
               last_obs::PeriodsSinceEpoch=Undated(typemin(Int)),
               order::Integer=1)

computes an unconditional forecast of the variables of the model

Keyword arguments

periods::Integer: number of forecasted periods [required]

forecast_mode::ForecastModes: one of histval or calibsmoother [required]

datafile::String: file with the observations for the smoother

first_obs::PeriodsSinceEpoch: first period used by smoother (default: first observation in the file)

first_period::PeriodsSinceEpoch: initial_period for the forecast (default when histval: Undated(0), default when calibsmoother: last period of the smoother)

last_obs::PeriodsSinceEpoch: last period used by smoother (default: last observation in the file)

order::Integer: order of local approximation

Dynare.recursive_forecasting! — Function

recursive_forecasting!(; Np::Integer,
                       first_period::PeriodsSinceEpoch=Undated(typemin(Int)), 
                       last_period::PeriodsSinceEpoch=Undated(typemax(Int)), 
                       context::Context=context,
                       datafile::String="", 
                       first_obs::PeriodsSinceEpoch=Undated(typemin(Int)), 
                       last_obs::PeriodsSinceEpoch=Undated(typemax(Int)), 
                       order::Integer=1)

computes an unconditional recursive forecast by adding one period to the sample used for the smoother before forecasting over Np periods.

Keyword arguments

Np::Integer: number of forecasted periods [required]

first_period::PeriodsSinceEpoch: initial period of first forecast [required]

last_period::PeriodsSinceEpoch: initial period of last forecast [required]

datafile::String: file with the observations for the smoother

first_obs::PeriodsSinceEpoch: first period used by smoother (default: first observation in the file)

last_obs::PeriodsSinceEpoch: last period used by smoother (default: last observation in the file)

order::Integer: order of local approximation